Adds email preview and confirmation to delivery resend
- Adds route and controller: `deliveries/show/resend-confirm`
- Moves resend action from resend into resend-confirm
- Moves action buttons out of panel footer for consistency
- Updates new delivery creation to show infoMessage after creating a delivery
