+++ app/router.js
@@ -11,6 +11,7 @@ Router.map(function() {
   this.route('deliveries', function() {
     this.route('show', { path: '/:transfer_id'}, function () {
       this.route('resend', {});
+      this.route('resend-confirm');
     });
     this.route('new', function() {
       this.route('select-project');
