+++ app/routes/deliveries/show/resend-confirm.js
@@ -1,0 +1,8 @@
+import Ember from 'ember';
+
+export default Ember.Route.extend({
+  setupController(controller, model) {
+    this._super(controller, model);
+    controller.generatePreview();
+  }
+});
