+++ app/templates/deliveries/show/resend-confirm.hbs
@@ -1,0 +1,18 @@
+{{delivery-breadcrumbs selectedRouteName="deliveries.show.resend" transfer=model}}
+<h3>Re-send Delivery Email</h3>
+<p class="body">
+  The email below will be sent when you click <strong>Resend Delivery</strong>. If you do not wish to send this email,
+  click <strong>Back</strong>
+</p>
+
+{{#if emailMessage}}
+  {{delivery-email emailMessage}}
+{{else}}
+  <div class="well well-sm">{{#loading-indicator}}Generating Preview{{/loading-indicator}}</div>
+{{/if}}
+
+{{error-message-alert errorMessages}}
+
+{{#bs-button type="default" onClick=(action 'back') class="back-button"}}Back{{/bs-button}}
+{{#bs-button type="primary" disabled=disableNext onClick=(action 'resend') class="pull-right next-button"}}Resend Delivery{{/bs-button}}
+
